chore: split router v1 from main.py
