+++ src/domain/handlers/job_processors.py
@@ -1,8 +1,5 @@
-from src.domain.services.job_service import JobService
-
-
 def process_filters(
     picker_id: str,
-    job_service: JobService
+    job_service
 ):
     job_service.process(int(picker_id))
